This feature adds the Grunt build process for component, extension, themes and menu type plugins.
The plugin system uses Bower to locate and install the files. All plugins must
be defined as AMD modules using the define function
i.e. define(name, deps, factory)

The following grunt tasks compile the less/css, handlebars into the build
directory:-

	grunt-contrib-less
    grunt-contrib-handlebars

Then there is a custom task to scan for plugins and bundle the into a
single amd module, this is grunt-requirejs-bundle. This is followed by
grunt-bower-requirejs, a third party task that inspects bower packages and
wires-up the packages main js file with requirejs.

Finally we can run grunt-contrib-requirejs to compile the prepared modules.