Added layout sort override
Added a snippet of code which allows you to override the left/right sort order of components by applying '_sortComponents': false to the container block.

Ideally used for when you have an MCQ component left aligned and Graphic component right aligned on desktop but want the Graphic component to appear above the MCQ in mobile view. 

All component JSON is unchanged, only amendment is to define the right aligned Graphic component above the left aligned MCQ component in the components.json.